requirements.txt
jupyter
pytest==7.1.3
pytest_pycodestyle
flask
pymodm
dnspython
requests
Pillow
